Déplace les exports pour faire de la place aux nouveaux tableaux
