+++ app/views/admin/exports/memoire/_photo.html.haml
