+++ app/views/admin/exports/memoire/_photos_table.html.haml
