+++ app/views/admin/exports/palissy/new.html.haml
@@ -1,6 +1,6 @@
 %main#contenu
   .fr-container.fr-mb-8w
-    = render "shared/breadcrumbs", links: [["Admin", admin_path], ["Exports par département", admin_exports_path]], current_page_label: "Export Palissy pour #{@departement}"
+    = render "shared/breadcrumbs", links: [["Admin", admin_path], ["Exports par département", admin_exports_pop_path]], current_page_label: "Export Palissy pour #{@departement}"
 
     %h1 Export Palissy pour #{@departement}
