+++ app/views/admin/exports/pop/index.html.haml
@@ -29,11 +29,11 @@
                 %div + #{departement.recensements.to_a.reject(&:pop_export_memoire_id).sum { _1.photos.reject(&:exportable?).length }} photos non exportables
               %td.co-white-space-nowrap
                 - departement.pop_exports.palissy.each do |pop_export|
-                  .fr-mb-2w= link_to pop_export, admin_palissy_export_path(pop_export), class: "fr-link"
-                %div= link_to "nouvel export Palissy …", new_admin_palissy_export_path(departement_code: departement.code), class: "fr-link"
+                  .fr-mb-2w= link_to pop_export, admin_exports_palissy_path(pop_export), class: "fr-link"
+                %div= link_to "nouvel export Palissy …", new_admin_exports_palissy_path(departement_code: departement.code), class: "fr-link"
                 - departement.pop_exports.memoire.each do |pop_export|
-                  .fr-mb-2w= link_to pop_export, admin_memoire_export_path(pop_export), class: "fr-link"
-                %div= link_to "nouvel export Mémoire…", new_admin_memoire_export_path(departement_code: departement.code), class: "fr-link"
+                  .fr-mb-2w= link_to pop_export, admin_exports_memoire_path(pop_export), class: "fr-link"
+                %div= link_to "nouvel export Mémoire…", new_admin_exports_memoire_path(departement_code: departement.code), class: "fr-link"
         %tfooter
           %tr
             %th Totaux
