+++ app/views/pages/admin.html.haml
@@ -25,4 +25,4 @@
             %li= link_to "Prévisualisation des emails", mail_previews_path
           %li= link_to "Suivi des tâches asynchrones (GoodJob)", "/good_job"
           %li= link_to "Codes de connexion", admin_session_codes_path
-          %li= link_to "Exports pour POP", admin_exports_path
+          %li= link_to "Exports pour POP", admin_exports_pop_path
